requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.3.0
httpx==0.24.1
GitPython==3.1.40
supabase==1.2.0